requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic
httpx
python-dotenv
authlib
